==> 10.Rerec.ipynb <==
%cd data
!python pca.py


!pip uninstall -y faiss faiss-cpu faiss-gpu faiss-gpu-cu11 faiss-gpu-cu12
!pip install --no-cache-dir "faiss-gpu-cu12>=1.8.0,<2.0.0"


%cd data
!python retrival_users.py

!pip uninstall -y tensorflow tensorflow-cpu tb-nightly tensorboard tensorboard-data-server tensorboard-plugin-wit
!pip install -U "numpy>=1.26.4" "tensorboard>=2.17.0"


!bash experiments/beauty.bash --dataset beauty
# print("")
